Extract class CreateTcpImposterTests
