+++ MbDotNet.Tests/Client/MountebankClientTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Linq;
 
-using MbDotNet.Enums;
 using MbDotNet.Interfaces;
 using MbDotNet.Models.Imposters;
 
@@ -15,7 +14,7 @@ namespace MbDotNet.Tests.Client
     {
         private IClient _client;
         private Mock<IRequestProxy> _mockRequestProxy;
-        
+
         [TestInitialize]
         public void TestInitialize()
         {
@@ -35,48 +34,6 @@ namespace MbDotNet.Tests.Client
         {
             this._client.CreateTcpImposter(123);
             Assert.AreEqual(0, this._client.Imposters.Count);
-        }
-
-        [TestMethod]
-        public void CreateTcpImposter_WithoutName_SetsNameToNull()
-        {
-            var imposter = this._client.CreateTcpImposter(123);
-
-            Assert.IsNotNull(imposter);
-            Assert.IsNull(imposter.Name);
-        }
-
-        [TestMethod]
-        public void CreateTcpImposter_WithName_SetsName()
-        {
-            const string expectedName = "Service";
-
-            var imposter = this._client.CreateTcpImposter(123, expectedName);
-            
-            Assert.IsNotNull(imposter);
-            Assert.AreEqual(expectedName, imposter.Name);
-        }
-
-        [TestMethod]
-        public void CreateTcpImposter_WithoutMode_SetsModeToText()
-        {
-            const string expectedMode = "text";
-
-            var imposter = this._client.CreateTcpImposter(123, null);
-            
-            Assert.IsNotNull(imposter);
-            Assert.AreEqual(expectedMode, imposter.Mode);
-        }
-
-        [TestMethod]
-        public void CreateTcpImposter_WithMode_SetsMode()
-        {
-            const string expectedMode = "binary";
-
-            var imposter = this._client.CreateTcpImposter(123, null, TcpMode.Binary);
-
-            Assert.IsNotNull(imposter);
-            Assert.AreEqual(expectedMode, imposter.Mode);
         }
 
         [TestMethod]
