+++ MbDotNet.Tests/MbDotNet.Tests.csproj
@@ -66,6 +66,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Client\CreateHttpImposterTests.cs" />
+    <Compile Include="Client\CreateTcpImposterTests.cs" />
     <Compile Include="Models\Imposters\HttpsImposterTests.cs" />
     <Compile Include="Models\Imposters\HttpImposterTests.cs" />
     <Compile Include="Models\Imposters\TcpImposterTests.cs" />
